Moving calculate date from module to utils
