+++ tests/unit/plugins/modules/test_vm_snapshot.py
@@ -5,7 +5,6 @@ __metaclass__ = type
 import sys
 
 import pytest
-import datetime
 
 from ansible_collections.scale_computing.hypercore.plugins.modules import vm_snapshot
 from ansible_collections.scale_computing.hypercore.plugins.module_utils.vm_snapshot import (
@@ -277,11 +276,6 @@ class TestEnsurePresent:
         vm_object = mocker.MagicMock(spec=VM)
         vm_object.uuid = "123"
 
-        # Mock check_parameters
-        mocker.patch(
-            "ansible_collections.scale_computing.hypercore.plugins.modules.vm_snapshot.check_parameters"
-        ).return_value = module
-
         # Mock send_create_request and task
         task = dict()
         mocker.patch(
@@ -379,52 +373,3 @@ class TestEnsureAbsent:
         print(result, expected_result)
         assert isinstance(result, tuple)
         assert result == expected_result
-
-
-# Test calculate_data() module function
-class TestCalculateDate:
-    @pytest.mark.parametrize(
-        "days, expected_output",
-        [
-            (1, (datetime.datetime.today().date() + datetime.timedelta(days=1))),
-            (0, None),
-            (None, None),
-        ],
-    )
-    def test_calculate_date(self, days, expected_output):
-        result = vm_snapshot.calculate_date(days)
-        if result:
-            result = result.date()
-        assert result == expected_output
-
-
-# Test check_parameters module function
-class TestCheckParameters:
-    @pytest.mark.parametrize(
-        "retain_for, expected_retain_for",
-        [
-            (1, datetime.datetime.today().date() + datetime.timedelta(days=1)),
-            (0, None),
-            (None, None),
-        ],
-    )
-    def test_check_parameters(self, create_module, retain_for, expected_retain_for):
-        module = create_module(
-            params=dict(
-                cluster_instance=dict(
-                    host="https://my.host.name", username="user", password="pass"
-                ),
-                vm_name="this-VM",
-                label="this-label",
-                retain_for=retain_for,
-                replication=True,
-                state="present",
-            )
-        )
-        expected_params = module.params.copy()
-        expected_params["retain_for"] = expected_retain_for
-
-        result = vm_snapshot.check_parameters(module)
-        if result.params["retain_for"]:
-            result.params["retain_for"] = result.params["retain_for"].date()
-        assert result.params == expected_params
